tests/wpt/web-platform-tests/battery-status/multiple-promises.https.html

<!DOCTYPE html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/test-only-api.js"></script>
<script src="resources/battery-status-helpers.js"></script>
</head>
<body>
<script>
battery_status_test(async (t, mockBatteryMonitor) => {
  const promise1 = navigator.getBattery();
  const promise2 = navigator.getBattery();
  assert_equals(promise1, promise2);

  mockBatteryMonitor.setBatteryStatus(false, 10, 20, 0.5);
  mockBatteryMonitor.verifyBatteryStatus(await promise1);
  mockBatteryMonitor.verifyBatteryStatus(await promise2);
}, 'multiple parallel invocations of navigator.getBattery()');
</script>
</body>
</html>
